(ios) Using SwapInWalletDetails when tapping on "incoming amount" from Home screen. Much refactoring was required to get everything working properly on iPad.
