+++ phoenix-ios/phoenix-ios/views/configuration/danger zone/drain wallet/DrainWalletView_Confirm.swift
@@ -196,7 +196,7 @@ struct DrainWalletView_Confirm: MVISubView {
 	}
 	
 	func popToWrapper(_ destination: PopToDestination) {
-		log.trace("popToWrapper()")
+		log.trace("popToWrapper(\(destination))")
 		
 		popToDestination = destination
 		popTo(destination)
