+++ phoenix-ios/phoenix-ios/views/environment/DeepLink.swift
@@ -78,3 +78,24 @@ class DeepLinkManager: ObservableObject {
 		}
 	}
 }
+
+enum PopToDestination: CustomStringConvertible {
+	case RootView(followedBy: DeepLink? = nil)
+	case ConfigurationView(followedBy: DeepLink? = nil)
+	
+	var followedBy: DeepLink? {
+		switch self {
+			case .RootView(let followedBy)          : return followedBy
+			case .ConfigurationView(let followedBy) : return followedBy
+		}
+	}
+	
+	public var description: String {
+		switch self {
+		case .RootView(let followedBy):
+			return "RootView(followedBy: \(followedBy?.rawValue ?? "nil")"
+		case .ConfigurationView(let followedBy):
+			return "ConfigurationView(follwedBy: \(followedBy?.rawValue ?? "nil")"
+		}
+	}
+}
