+++ phoenix-ios/phoenix-ios/views/receive/ReceiveLightningView.swift
@@ -104,10 +104,9 @@ struct ReceiveLightningView: View {
 		.onAppear {
 			onAppear()
 		}
-		.navigationStackDestination( // For iOS 16+
-			isPresented: $currencyConverterOpen,
-			destination: currencyConverterView
-		)
+		.navigationStackDestination(isPresented: $currencyConverterOpen) { // For iOS 16+
+			currencyConverterView()
+		}
 		.onChange(of: mvi.model) { newModel in
 			onModelChange(model: newModel)
 		}
